pyproject.toml
[project]
name = "mcp-resk"
version = "0.1.0"
description = "RESK-MCP is an open-source Python library designed to add a robust security and management layer over the official MCP Library."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "mcp[cli]>=1.9.0",
    "pydantic>=2.11.4",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "slowapi>=0.1.9",
    "types-pyyaml>=6.0.12.20250516",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "flake8>=7.2.0",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "types-pyyaml>=6.0.12.20250516",
]
